Refactor: 어드민 Header 컴포넌트로 분리
