+++ app/(kahlua)/admin/page.tsx
@@ -1,4 +1,5 @@
 'use client';
+import Header from '@/components/admin/Header';
 import AdminPageButton from '@/components/admin/adminPageButton';
 
 const LEFT_COLUMN_ITEMS = [
@@ -51,18 +52,5 @@ const page = () => {
     </div>
   );
 };
-
-const Header = () => (
-  <div className="w-full h-[148px] pad:h-[260px] bg-gray-5 py-8 pad:py-16 flex flex-col justify-center items-center px-4">
-    <div className="w-full pad:w-[786px] dt:w-[1200px] h-full flex flex-col gap-6">
-      <p className="text-[36px] pad:text-[64px] font-semibold leading-none">
-        Admin
-      </p>
-      <p className="text-[16px] pad:text-[24px] font-semibold">
-        KAHLUA Admin Page
-      </p>
-    </div>
-  </div>
-);
 
 export default page;
