+++ components/admin/message/MessageTable.tsx
@@ -1,0 +1,4 @@
+const MessageTable = () => {
+  return <></>;
+};
+export default MessageTable;
